fixed evaluation bugs

+++ gym_mapf/mapf_env/mapf_env.cpp
@@ -392,8 +392,8 @@ MultiAgentState *MapfEnv::reset() {
 }
 
 MultiAgentState *MapfEnv::locations_to_state(const vector<Location> &locations) {
-    int mul = 1;
-    int sum = 0;
+    uint64_t mul = 1;
+    uint64_t sum = 0;
     int n_options = this->grid->id_to_loc.size();
 
     sum += locations[0].id * mul;
